fix : coupon 상태 분리
