+++ src/basic/models/coupon/useCoupon.ts
@@ -1,0 +1,55 @@
+import { useCallback, useState } from "react";
+import { Coupon } from "../../../types";
+import { initialCoupons } from "./constants";
+import { NotificationType } from "../../hooks/useNotifications";
+
+export const useCoupon = (
+  addNotification: (message: string, type: NotificationType) => void
+) => {
+  const [coupons, setCoupons] = useState<Coupon[]>(() => {
+    const saved = localStorage.getItem("coupons");
+    if (saved) {
+      try {
+        return JSON.parse(saved);
+      } catch {
+        return initialCoupons;
+      }
+    }
+    return initialCoupons;
+  });
+
+  const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
+
+  const addCoupon = useCallback(
+    (newCoupon: Coupon) => {
+      const existingCoupon = coupons.find((c) => c.code === newCoupon.code);
+      if (existingCoupon) {
+        addNotification("이미 존재하는 쿠폰 코드입니다.", "error");
+        return;
+      }
+      setCoupons((prev) => [...prev, newCoupon]);
+      addNotification("쿠폰이 추가되었습니다.", "success");
+    },
+    [coupons, addNotification]
+  );
+
+  const deleteCoupon = useCallback(
+    (couponCode: string) => {
+      setCoupons((prev) => prev.filter((c) => c.code !== couponCode));
+      if (selectedCoupon?.code === couponCode) {
+        setSelectedCoupon(null);
+      }
+      addNotification("쿠폰이 삭제되었습니다.", "success");
+    },
+    [selectedCoupon, addNotification]
+  );
+
+  return {
+    coupons,
+    setCoupons,
+    addCoupon,
+    deleteCoupon,
+    selectedCoupon,
+    setSelectedCoupon,
+  };
+};
